requirements.txt
uvicorn
fastapi
python-dotenv
motor
python-dateutil
pytz
python-jose
passlib
cryptography
python-multipart
bcrypt
requests
librouteros